mac: do not insert Homebrew LLVM in PATH
Avoid polluting the PATH variable with a Homebrew LLVM/Clang, and prefer
official Xcode/Command Line Tools installations instead.

The motivation for having Homebrew LLVM in PATH was to obtain access to
`clangd` as an LSP server. This can instead be achieved through
alternate means, such as a local symlink in `~/bin`.